Fix #76
Add a new ThenSetPropertyUsing() method at ColumnInfo in order to allow SetProp() chaining.
